Add automated testing
