+++ Part2/MARKING.md
@@ -1,0 +1,10 @@
+# Automated test guidance
+
+The automated tests expect the following:
+
+- Client recieves "Hello Client!" from server and prints it to the console
+- Server recieves "Hello Server!" from client and prints it to the console
+- Both client and server terminate after sending and recieving messages
+- Both client and server terminate within 10 seconds
+
+The tests ignore whitespace and capitalisation, and any other output.
