+++ Part3/MARKING.md
@@ -1,0 +1,8 @@
+# Automated test guidance
+
+The automated tests expect the following:
+
+- Client prompts user for username
+- Afterwards, client prompts user for messages which are sent to each fellow client via the server
+
+The test ignores any formatting, it only checks that the input of one client appears in the output of another client.
